refactor: creat getUpdateChoices.ts
